+++ cmd/mdmctl/get.go
@@ -102,7 +102,7 @@ Examples:
   mdmctl get devices
 
   # Get a device by serial (TODO implement filtering)
-  mdmctl get devices -serial=C02ABCDEF
+  mdmctl get devices -serials=C02ABCDEF
 `
 	fmt.Println(getUsage)
 	return nil
@@ -121,7 +121,7 @@ func (out *devicesTableOutput) BasicFooter() {
 func (cmd *getCommand) getDevices(args []string) error {
 	flagset := flag.NewFlagSet("devices", flag.ExitOnError)
 	var (
-		flFilterSerials = flagset.String("serials", "", "comma separated list of serials to search")
+		flFilterSerials = flagset.String("serials", "", "device serial, optionally comma-separated")
 	)
 	flagset.Usage = usageFor(flagset, "mdmctl get devices [flags]")
 	if err := flagset.Parse(args); err != nil {