+++ cmd/mdmctl/remove_blueprints.go
@@ -10,7 +10,7 @@ import (
 func (cmd *removeCommand) removeBlueprints(args []string) error {
 	flagset := flag.NewFlagSet("remove-blueprints", flag.ExitOnError)
 	var (
-		flBlueprintName = flagset.String("name", "", "name of blueprint, optionally comma separated")
+		flBlueprintName = flagset.String("name", "", "name of blueprint, optionally comma-separated")
 	)
 	flagset.Usage = usageFor(flagset, "mdmctl remove blueprints [flags]")
 	if err := flagset.Parse(args); err != nil {