+++ cmd/mdmctl/remove_devices.go
@@ -14,15 +14,15 @@ import (
 func (cmd *removeCommand) removeDevices(args []string) error {
 	flagset := flag.NewFlagSet("remove-devices", flag.ExitOnError)
 	var (
-		flIdentifier = flagset.String("udid", "", "device UDID, optionally comma separated")
-		flSerial     = flagset.String("serial", "", "device serial, optionally comma separated")
+		flIdentifier = flagset.String("udid", "", "device UDID, optionally comma-separated")
+		flSerials    = flagset.String("serials", "", "device serial, optionally comma-separated")
 	)
 	flagset.Usage = usageFor(flagset, "mdmctl remove devices [flags]")
 	if err := flagset.Parse(args); err != nil {
 		return err
 	}
 
-	if *flIdentifier == "" && *flSerial == "" {
+	if *flIdentifier == "" && *flSerials == "" {
 		flagset.Usage()
 		return errors.New("bad input: device UDID or Serial must be provided")
 	}
@@ -31,8 +31,8 @@ func (cmd *removeCommand) removeDevices(args []string) error {
 	if *flIdentifier != "" {
 		opts.UDIDs = strings.Split(*flIdentifier, ",")
 	}
-	if *flSerial != "" {
-		opts.Serials = strings.Split(*flSerial, ",")
+	if *flSerials != "" {
+		opts.Serials = strings.Split(*flSerials, ",")
 	}
 
 	ctx := context.Background()